chore: add evaluation compatibility shim
