+++ src/production/evolution/evomerge/integration_tests.py
@@ -4,9 +4,9 @@ import unittest
 from transformers import AutoModelForCausalLM, AutoTokenizer
 
 from .config import create_default_config
-from .evaluation import evaluate_model
 from .evolutionary_tournament import run_evolutionary_tournament
 from .utils import clean_up_models, generate_text
+from agent_forge.evaluation.evaluator import evaluate_model
 
 
 class TestEvoMergeIntegration(unittest.TestCase):
