+++ src/production/evolution/evomerge/merging/merger.py
@@ -17,11 +17,11 @@ from ..utils import (
     EvoMergeException,
     check_system_resources,
     clean_up_models,
-    evaluate_model,
     load_models,
     mask_model_weights,
     save_model,
 )
+from agent_forge.evaluation.evaluator import evaluate_model
 from .merge_techniques import MERGE_TECHNIQUES
 
 logger = logging.getLogger(__name__)
